requirements.txt
# discord==0.16.12
discord
aiohttp
pyyaml
requests
pyquery
youtube_dl
opuslib
GitPython
#required discord.py==0.16.12
